+++ tests/testthat/unit/test-key_by.R
@@ -3,14 +3,14 @@ context("  Key by")
 test_that("The key_by verb works as expected.", {
     expected <- DT[, .(ans = mean(mpg)), keyby = vs]
 
-    ans <- DT %>% start_expr %>% select(ans = mean(mpg)) %>% key_by(vs) %>% end_expr
+    ans <- DT %>% start_expr %>% transmute(ans = mean(mpg)) %>% key_by(vs) %>% end_expr
     expect_identical(ans, expected)
 
     expected <- DT[, .(mean = mean(mpg), sd = sd(mpg)), keyby = .(vs, am)]
 
     ans <- DT %>%
         start_expr %>%
-        select(mean = mean(mpg), sd = sd(mpg)) %>%
+        transmute(mean = mean(mpg), sd = sd(mpg)) %>%
         key_by(vs, am) %>%
         end_expr
 
@@ -18,7 +18,7 @@ test_that("The key_by verb works as expected.", {
 
     ans <- DT %>%
         start_expr %>%
-        select(mean = mean(mpg), sd = sd(mpg)) %>%
+        transmute(mean = mean(mpg), sd = sd(mpg)) %>%
         key_by("vs", "am", .parse = TRUE) %>%
         end_expr
 
@@ -32,5 +32,5 @@ test_that("The key_by verb works as expected.", {
         key_by(!!desired_group, "am", .parse = TRUE) %>%
         end_expr
 
-    expect_identical(ans, expected)
+    expect_equal(ans, expected)
 })