+++ NEWS.md
@@ -1,9 +1,9 @@
 # table.express 0.1.1.9000
 
 - Added joining verbs and a corresponding vignette describing them.
-- Added `.enlist` parameter to `select` for cases when the expression should not be wrapped in a
-  call to `list`.
-- Minor bug fixes in `select` for unnamed arguments or a named call involving `:`.
+- `transmute` is no longer an alias for `select`. This is to make the latter more flexible in some
+  bare selection cases (like combining several `tidyselect` calls), and leave the former more simple
+  and preferable for actual transmutation cases.
 
 # table.express 0.1.1