+++ NEWS.md
@@ -11,6 +11,8 @@
 - Added some debugging statements when option `table.express.verbose` is `TRUE`, or
   `.verbose = TRUE` in `start_expr`/`chain`.
 - Minor updates to main vignette.
+- Semi-breaking change: `*_sd` verbs now have `.SDcols` before their ellipsis. This should only
+  require code adjustment for `transmute_sd` when `.SDcols` was left missing.
 
 Special thanks to @leungi for checking consistency with `dplyr` in several cases.